Fix build exclusions in Makefile.
This became apparent when a zipped build still included the `build` directory. While the directory should always be empty (the default `make` task clears the build folder before zipping), it does include the .git folder, which will become larger as the repository grows.

At current revision, the reduction in the build size is about 80%:

    With .git folder:    putioroku.zip - 325kb
    Without .git folder: putioroku.zip - 65kb